requirements.txt
python-bale-bot
Pillow